pyproject.toml
[project]
name = "moku-go"
version = "0.1.0"
description = "CLI interface for Liquid Instruments Moku-Go device"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "moku>=3.3.3",
    "zeroconf>=0.147.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "hatchling>=1.21.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
moku-go = "moku_go.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["moku_go"]
